+++ arduino/filter_custom_biquad/filter_custom_biquad.ino
@@ -6,7 +6,7 @@ AudioInfo info(44100, 2, 16);
 AudioBoardStream lyrat(LyratV43);
 
 //Filtered Stream
-FilteredStream<int16_t, double> filtered(lyrat, info.channels);
+FilteredStream<int16_t, float> filtered(lyrat, info.channels);
 StreamCopy filter(filtered, lyrat);
 
 //CsvOutput
@@ -14,16 +14,16 @@ CsvOutput<int16_t> out(Serial);
 StreamCopy csv(out, filtered);
 
 //Filter Coefficients
-const double b_0 = 10;
-const double b_1 = 0;
-const double b_2 = 0;
+const float b_0 = 10;
+const float b_1 = 0;
+const float b_2 = 0;
 
-const double a_0 = 1;
-const double a_1 = 0;
-const double a_2 = 0;
+const float a_0 = 1;
+const float a_1 = 0;
+const float a_2 = 0;
 
-const double b_coefficients[] = { b_0, b_1, b_2};
-const double a_coefficients[] = { a_0, a_1, a_2};
+const float b_coefficients[] = { b_0, b_1, b_2};
+const float a_coefficients[] = { a_0, a_1, a_2};
 
 //Arduino Setup
 void setup() {
@@ -41,8 +41,8 @@ void setup() {
 
 
     //setup Filters for both Channels
-    filtered.setFilter(0, new BiQuadDF1<double>(b_coefficients, a_coefficients));
-    filtered.setFilter(1, new BiQuadDF1<double>(b_coefficients, a_coefficients));
+    filtered.setFilter(0, new BiQuadDF1<float>(b_coefficients, a_coefficients));
+    filtered.setFilter(1, new BiQuadDF1<float>(b_coefficients, a_coefficients));
 
     Serial.println("READY");
 }
